+++ mitra_adapters/src/dynamic_config.rs
@@ -4,6 +4,7 @@ use serde_json::{Value as JsonValue};
 use mitra_models::{
     database::{DatabaseClient, DatabaseError, DatabaseTypeError},
     properties::constants::{
+        FAVORITE_EMOJIS,
         FEDERATED_TIMELINE_RESTRICTED,
         FILTER_BLOCKLIST_PUBLIC,
         FILTER_KEYWORDS,
@@ -14,19 +15,22 @@ use mitra_models::{
     },
 };
 use mitra_validators::errors::ValidationError;
+use mitra_utils::unicode::is_single_character;
 
 // Dynamic configuration parameters
-pub const EDITABLE_PROPERTIES: [&str; 4] = [
+pub const EDITABLE_PROPERTIES: [&str; 5] = [
     FEDERATED_TIMELINE_RESTRICTED,
     FILTER_BLOCKLIST_PUBLIC,
     FILTER_KEYWORDS,
     LIKE_EMOJI,
+    FAVORITE_EMOJIS,
 ];
 
 const LIKE_EMOJI_VARIANTS: [&str; 2] = [
     "thumbs_up",
     "heart",
 ];
+const DEFAULT_FAVORITE_EMOJIS: [&str; 8] = ["❤️", "😆", "🤔", "😢", "😡", "🎉", "💯", "👀"];
 
 pub fn validate_editable_parameter(
     name: &str,
@@ -34,6 +38,13 @@ pub fn validate_editable_parameter(
 ) -> Result<(), ValidationError> {
     let value = value.clone();
     match name {
+        FAVORITE_EMOJIS => {
+            let emojis: Vec<String> = serde_json::from_value(value)
+                .map_err(|_| ValidationError("invalid value type"))?;
+            if !emojis.iter().all(|text| is_single_character(text)) {
+                return Err(ValidationError("invalid unicode emoji"));
+            };
+        },
         FEDERATED_TIMELINE_RESTRICTED
             | FILTER_BLOCKLIST_PUBLIC =>
         {
@@ -59,6 +70,7 @@ pub fn validate_editable_parameter(
 #[derive(Deserialize, Serialize)]
 #[serde(default)]
 pub struct DynamicConfig {
+    pub favorite_emojis: Vec<String>,
     pub federated_timeline_restricted: bool,
     pub filter_blocklist_public: bool,
     pub filter_keywords: Vec<String>,
@@ -68,6 +80,9 @@ pub struct DynamicConfig {
 impl Default for DynamicConfig {
     fn default() -> Self {
         Self {
+            favorite_emojis: DEFAULT_FAVORITE_EMOJIS.iter()
+                .map(|emoji| emoji.to_string())
+                .collect(),
             federated_timeline_restricted: false,
             filter_blocklist_public: false,
             filter_keywords: vec![],