+++ mitra_api/src/mastodon_api/instance/types.rs
@@ -411,6 +411,7 @@ pub struct InstanceInfoV2 {
     login_message: String,
     new_accounts_read_only: bool,
     like_emoji: String,
+    favorite_emojis: Vec<String>,
     blockchains: Vec<BlockchainInfo>,
     ipfs_gateway_url: Option<String>,
 
@@ -497,6 +498,7 @@ impl InstanceInfoV2 {
             new_accounts_read_only:
                 matches!(config.registration.default_role, DefaultRole::ReadOnlyUser),
             like_emoji: dynamic_config.like_emoji,
+            favorite_emojis: dynamic_config.favorite_emojis,
             blockchains: config.blockchains().iter()
                 .map(BlockchainInfo::from)
                 .collect(),