+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added option to generate `Bite` activities to `create-activity` command.
 - Added `all_reactions` parameter to `/api/v1/favourites` endpoint.
 - Added `like_emoji` parameter to dynamic configuration.
+- Added `favorite_emojis` parameter to dynamic configuration.
 
 ### Changed